feat: add sign out function to account menu
A new async sign out function is created and imported into the account menu component. A new mutation is added to the AccountMenu component, using the new sign out function. On success of sign out, the user is redirected to the sign-in page. An onClick function has been added to the sign-out button to execute this functionality.