fix: adjust order total to display in cents
The commit modifies how the total order amount is displayed on the order summary table. Previously, it was shown in whole numbers, but after this change, the total will be displayed in cents (divided by 100), giving a more precise value. This modification aligns with the standard currency format in Brazil.